requirements.txt
python-telegram-bot>=20.0
yt-dlp>=2025.6.1
